fix(fish): pokemon color scripts greeting
krabby was not found during the execution of the fish_greeting
since the binary is no longer on the PATH since I disabled mise shims

Now, we just run krabby via mise and fallback to calling the
binary directly if mise is not available.

Also, removed the mise shims from the PATH since mise
auto removes shims if you use `mise activate` since
shims and mise's PATH are mutually exclusive.