fix(ghostty): tmux startup command
If the tmux server was not running the startup command would fail
because ghostty does not do any shell command parsing.

This ment that our `||` was being treated like a string instead of `OR`
which broke our error handling

So now, we wrap our tmux startup command with bash to ensure that it's
parsed as proper shell script and not a string

Also made tmux open to the `main` session by default if it exists, else
create it. This fixes the issue where a random session would be attached
to on launch. (idk if was random or just the last used session)

but it's all fixed so i'm happy