fix(ghostty): tmux startup again
Turns out my last fix didn't account for the fact that when tmux is run
as a one off command in bash, it does not have access to the rest of
my env.

So tmux would indeed start properly, but my XDG config's would not
exist in tmux's eyes, so my config reload hotkey would error out.
But the biggest breakage was the fact that it broke tmux-resurrect and
tmux-continuum, which was an absolute deal breaker

So, the solution was to move the tmux startup logic into my shell config
Now, it all works as intended, although not as elegant as I had it
before, but it all works, so no complaints here