feat(bin): create `tmux-autostart`
This is a bash script that runs tmux on supported terminals

This is supposed to be dropped in your shell config to auto
start tmux in the main session

Also created `02-tmux.fish` to run `tmux-autostart` right
after loading all env's to maximize speed while keeping it
portable.

If we want to run tmux before loading env, we would have to
use the path of the tmux bin to call it, which would make this
not portable across systems since the tmux location could
and probably will differ.